Fix iOS metadata compilation: move NSObject delegate to target-specific source sets
The metadata compiler for iosMain cannot resolve NSObject as a supertype (ObjC
interop class hierarchies aren't available in metadata mode). Restructure so:

- iosMain/IosDownloadHelper.kt: expect fun nsUrlSessionDownload(...) using only
  Kotlin types — compiles cleanly in metadata mode.
- iosArm64Main/IosDownloadHelper.kt and iosSimulatorArm64Main/IosDownloadHelper.kt:
  actual fun using NSObject + NSURLSessionDownloadDelegateProtocol, only compiled
  by the full target toolchains where Foundation is available.
- PlatformDbSupport.ios.kt: actual downloadFileToPath delegates to nsUrlSessionDownload
  with lambda wrappers, no ObjC supertypes in iosMain.
- Remove IosBackgroundDownloadDelegate.kt (delegate is now inlined as anonymous
  object inside the target-specific actual fun).
