+++ composeApp/src/iosArm64Main/kotlin/com/slovy/slovymovyapp/data/remote/IosDownloadHelper.kt
@@ -1,0 +1,98 @@
+package com.slovy.slovymovyapp.data.remote
+
+import kotlinx.cinterop.ExperimentalForeignApi
+import kotlinx.coroutines.suspendCancellableCoroutine
+import kotlinx.io.files.Path
+import platform.Foundation.*
+import kotlin.coroutines.resume
+import kotlin.coroutines.resumeWithException
+
+@OptIn(ExperimentalForeignApi::class)
+internal actual suspend fun nsUrlSessionDownload(
+    url: String,
+    headers: Map<String, String>,
+    destPath: Path,
+    tempPath: Path,
+    onProgress: (DownloadProgress) -> Unit,
+    cancelToken: CancelToken,
+    moveFile: (from: Path, to: Path) -> Boolean,
+    deleteFile: (path: Path) -> Unit,
+) = suspendCancellableCoroutine<Unit> { cont ->
+    val sessionId = "com.slovy.slovymovyapp.dl.${
+        NSURL.fileURLWithPath(destPath.toString()).lastPathComponent ?: destPath.name
+    }"
+    val config = NSURLSessionConfiguration.backgroundSessionConfigurationWithIdentifier(sessionId)
+    config.discretionary = false
+    config.allowsCellularAccess = true
+
+    val delegate = object : NSObject(), NSURLSessionDownloadDelegateProtocol {
+        override fun URLSession(
+            session: NSURLSession,
+            downloadTask: NSURLSessionDownloadTask,
+            didWriteData: Long,
+            totalBytesWritten: Long,
+            totalBytesExpectedToWrite: Long,
+        ) {
+            if (cancelToken.isCancelled) {
+                downloadTask.cancel()
+                return
+            }
+            val total = if (totalBytesExpectedToWrite > 0) totalBytesExpectedToWrite else null
+            onProgress(DownloadProgress(totalBytesWritten, total))
+        }
+
+        override fun URLSession(
+            session: NSURLSession,
+            downloadTask: NSURLSessionDownloadTask,
+            didFinishDownloadingToURL: NSURL,
+        ) {
+            // The file at didFinishDownloadingToURL is temporary — iOS deletes it when
+            // this method returns. Move it synchronously before returning.
+            val srcPath = Path(
+                didFinishDownloadingToURL.path ?: run {
+                    session.finishTasksAndInvalidate()
+                    cont.resumeWithException(IllegalStateException("Nil download location URL"))
+                    return
+                }
+            )
+            val moved = moveFile(srcPath, tempPath) && moveFile(tempPath, destPath)
+            session.finishTasksAndInvalidate()
+            if (moved) {
+                cont.resume(Unit)
+            } else {
+                cont.resumeWithException(
+                    IllegalStateException("Failed to move downloaded file to destination")
+                )
+            }
+        }
+
+        override fun URLSession(
+            session: NSURLSession,
+            task: NSURLSessionTask,
+            didCompleteWithError: NSError?,
+        ) {
+            if (didCompleteWithError != null && cont.isActive) {
+                deleteFile(tempPath)
+                cont.resumeWithException(
+                    IllegalStateException("Download failed: ${didCompleteWithError.localizedDescription}")
+                )
+            }
+        }
+    }
+
+    val session = NSURLSession.sessionWithConfiguration(config, delegate = delegate, delegateQueue = null)
+    val nsUrl = NSURL.URLWithString(url) ?: run {
+        cont.resumeWithException(IllegalArgumentException("Invalid URL: $url"))
+        return@suspendCancellableCoroutine
+    }
+    val request = NSMutableURLRequest.requestWithURL(nsUrl)
+    headers.forEach { (k, v) -> request.setValue(v, forHTTPHeaderField = k) }
+    val task = session.downloadTaskWithRequest(request)
+
+    cont.invokeOnCancellation {
+        task.cancel()
+        session.invalidateAndCancel()
+        deleteFile(tempPath)
+    }
+    task.resume()
+}
